requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
Flask-Login==0.5.0
stripe==2.58.0
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2-binary==2.9.1 